+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/models/assessRisksAndNeeds/ArnViolencePredictorScore.kt
@@ -1,10 +1,11 @@
 package uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.assessRisksAndNeeds
 
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.hmpps.ViolencePredictor
+import java.math.BigDecimal
 
 data class ArnViolencePredictorScore(
   val ovpRisk: String? = null,
-  val twoYears: Int? = null,
+  val twoYears: BigDecimal? = null,
 ) {
   fun toViolencePredictor(useV2NumericalValue: Boolean = true): ViolencePredictor =
     ViolencePredictor(
